+++ scripts/checks.sh
@@ -21,9 +21,8 @@ cargo package -p cloud-sdk-hetzner --allow-dirty \
 cargo package -p cloud-sdk-hetzner-reqwest --allow-dirty \
     --config 'patch.crates-io.cloud-sdk.path="crates/cloud-sdk"' \
     --config 'patch.crates-io.cloud-sdk-hetzner.path="crates/cloud-sdk-hetzner"'
-cargo package -p cloud-sdk-hetzner-sanitization --allow-dirty \
-    --config 'patch.crates-io.cloud-sdk.path="crates/cloud-sdk"' \
-    --config 'patch.crates-io.cloud-sdk-hetzner.path="crates/cloud-sdk-hetzner"'
+cargo package -p cloud-sdk-sanitization --allow-dirty \
+    --config 'patch.crates-io.cloud-sdk.path="crates/cloud-sdk"'
 cargo package -p cloud-sdk-hetzner-testkit --allow-dirty \
     --config 'patch.crates-io.cloud-sdk.path="crates/cloud-sdk"' \
     --config 'patch.crates-io.cloud-sdk-hetzner.path="crates/cloud-sdk-hetzner"'